refactor(accounts): split account operations
