+++ src/components/KiloCodeExportDialog.tsx
@@ -22,7 +22,7 @@ import AddTokenDialog from "~/features/TokenProvisioning/components/AddTokenDial
 import { buildDefaultTokenCreatePrefill } from "~/features/TokenProvisioning/components/AddTokenDialog/defaultTokenCreatePrefill"
 import { useAccountData } from "~/hooks/useAccountData"
 import { useSafeExportAction } from "~/hooks/useSafeExportAction"
-import { ensureAccountApiToken } from "~/services/accounts/accountOperations"
+import { ensureAccountApiToken } from "~/services/accounts/ensureAccountApiToken"
 import {
   resolveDefaultTokenQuickCreateResolution,
   TOKEN_QUICK_CREATE_RESOLUTION_KINDS,
