+++ src/components/dialogs/ChannelDialog/hooks/useChannelDialog.ts
@@ -10,10 +10,10 @@ import {
 } from "~/components/dialogs/ChannelDialog/utils/advisoryWarning"
 import { DIALOG_MODES, type DialogMode } from "~/constants/dialogModes"
 import { SITE_TYPES } from "~/constants/siteType"
-import { ensureAccountApiToken } from "~/services/accounts/accountOperations"
 import { selectSingleNewApiTokenByIdDiff } from "~/services/accounts/accountPostSaveWorkflow"
 import { accountStorage } from "~/services/accounts/accountStorage"
 import { createUnsupportedTodayStatsAvailability } from "~/services/accounts/accountTodayStats"
+import { ensureAccountApiToken } from "~/services/accounts/ensureAccountApiToken"
 import {
   resolveDefaultTokenQuickCreateResolution,
   TOKEN_QUICK_CREATE_RESOLUTION_KINDS,
