+++ tests/features/AccountManagement/hooks/useAccountDialog.currentTabDetection.test.tsx
@@ -69,15 +69,21 @@ vi.mock("~/contexts/UserPreferencesContext", async (importOriginal) => {
   }
 })
 
-vi.mock("~/services/accounts/accountOperations", async (importOriginal) => {
+vi.mock("~/services/accounts/accountAutoDetection", () => ({
+  autoDetectAccount: mockAutoDetectAccount,
+}))
+
+vi.mock("~/services/accounts/siteName", () => ({
+  getSiteName: mockGetSiteName,
+}))
+
+vi.mock("~/services/accounts/accountPersistence", async (importOriginal) => {
   const actual =
     await importOriginal<
-      typeof import("~/services/accounts/accountOperations")
+      typeof import("~/services/accounts/accountPersistence")
     >()
   return {
     ...actual,
-    autoDetectAccount: mockAutoDetectAccount,
-    getSiteName: mockGetSiteName,
     validateAndSaveAccount: mockValidateAndSaveAccount,
   }
 })
