+++ tests/features/AccountManagement/hooks/useAccountDialog.redetectPreservesCustomData.test.tsx
@@ -72,10 +72,10 @@ vi.mock("~/services/protectionBypass/client", () => ({
     }),
 }))
 
-vi.mock("~/services/accounts/accountOperations", async (importOriginal) => {
+vi.mock("~/services/accounts/accountAutoDetection", async (importOriginal) => {
   const actual =
     await importOriginal<
-      typeof import("~/services/accounts/accountOperations")
+      typeof import("~/services/accounts/accountAutoDetection")
     >()
   return {
     ...actual,
