+++ tests/services/accountPersistence.autoProvisionKeyOnAccountAdd.test.ts
@@ -3,7 +3,7 @@ import { beforeEach, describe, expect, it, vi } from "vitest"
 import { Storage } from "@plasmohq/storage"
 
 import { OPENROUTER_WEB_ORIGIN, SITE_TYPES } from "~/constants/siteType"
-import { validateAndSaveAccount } from "~/services/accounts/accountOperations"
+import { validateAndSaveAccount } from "~/services/accounts/accountPersistence"
 import { accountStorage } from "~/services/accounts/accountStorage"
 import { DefaultTokenLifecyclePolicyBlockedError } from "~/services/accounts/defaultTokenLifecycle"
 import { TOKEN_PROVISIONING_BLOCK_REASONS } from "~/services/apiAdapters/contracts/tokenProvisioning"
@@ -86,7 +86,7 @@ const CHECK_IN_DISABLED = buildCheckInConfig({
 
 const flushPromises = () => new Promise((resolve) => setTimeout(resolve, 0))
 
-describe("accountOperations auto-provision key on add", () => {
+describe("accountPersistence auto-provision key on add", () => {
   beforeEach(async () => {
     fetchAccountDataMock.mockReset()
     ensureDefaultApiTokenForAccountMock.mockReset()
